pyproject.toml
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/__pycache__/*", "*/venv/*", "*/.venv/*"]

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "invoicegen"
version = "1.0.0"
description = "Synthetic invoice generation and LayoutLMv3 training pipeline"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "InvoiceGen Team", email = "team@invoicegen.com"}
]
keywords = ["invoice", "document-understanding", "layoutlmv3", "ocr", "ner"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "opencv-python>=4.5.0",
    "jinja2>=3.0.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "seqeval>=1.2.2",
    "scikit-learn>=1.2.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "isort>=5.12.0",
]
ocr = [
    "paddleocr>=2.6.0",
    "paddlepaddle>=2.4.0",
    "pytesseract>=0.3.10",
    "easyocr>=1.6.0",
]
all = [
    "invoicegen[dev,ocr]",
]

[project.urls]
Homepage = "https://github.com/yourusername/InvoiceGen"
Documentation = "https://github.com/yourusername/InvoiceGen/docs"
Repository = "https://github.com/yourusername/InvoiceGen"
"Bug Tracker" = "https://github.com/yourusername/InvoiceGen/issues"

[project.scripts]
invoicegen-build = "scripts.build_training_set:main"
invoicegen-train = "scripts.run_training:main"
invoicegen-serve = "deployment.api:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
